+++ src/routes/Interest/intereststock/chart/InterestStockDetailChartPage.jsx
@@ -8,19 +8,17 @@ import {
   fetchHankookStockBalance,
   fetchHankookStockCurrent,
 } from "../../../../lib/apis/hankookApi.jsx";
-import { Link, useNavigate, useParams } from "react-router-dom";
-import party, { fetchPartyInfo } from "../../../../lib/apis/party.jsx";
+import { useNavigate, useParams } from "react-router-dom";
+import { fetchPartyInfo } from "../../../../lib/apis/party.jsx";
 import { getApproval, regist } from "../../../../lib/apis/interest.jsx";
 import { SyncLoader } from "react-spinners";
-import ApproveInterestModal from "../../../../components/common/modal/ApproveInterestModal.jsx";
+import TradeButton from "../../../../components/common/button/TradeButton.jsx";
 
 export default function InterestStockDetailChartPage() {
   const partyKey = useParams().partyKey;
   const stockKey = useParams().stockKey;
   const [stockInfo, setStockInfo] = useState([]);
   const [stockBalance, setStockBalance] = useState([]);
-  const [isInterestStock, setIsInterestStock] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false); // 모달 열기 여부 상태 추가
 
   const navigate = useNavigate();
 
@@ -37,7 +35,6 @@ export default function InterestStockDetailChartPage() {
         const isActive =
           mystock.find((data) => data.stockKey === stockKey) !== undefined;
 
-        setIsInterestStock(isActive);
         setStockInfo(stockInfo);
         setStockBalance({
           data: fetchData,
@@ -50,22 +47,6 @@ export default function InterestStockDetailChartPage() {
     }
     fetchData();
   }, []);
-
-  const handleAddToInterestStock = async () => {
-    try {
-      console.log("Stock added to interest stock!");
-      const reqBody = {
-        stockKey: stockKey,
-      };
-
-      await regist(partyKey, reqBody);
-      setIsModalOpen(true); // 모달 열기
-      setIsInterestStock(true);
-    } catch (error) {
-      console.error(error);
-      throw Error(error);
-    }
-  };
 
   const handleAskingPriceButtonClick = () => {
     navigate(`/stockDetail/${partyKey}/${stockKey}/askingPrice`, {
@@ -148,37 +129,11 @@ export default function InterestStockDetailChartPage() {
             </Row>
             <StockDataFetcher stockBalance={stockBalance} />
 
-            {stockBalance.data ? (
-              <Row className="stock-detail-transaction-button">
-                <Button className="stock-detail-sell-button">
-                  <div className="stock-detail-sell-text">판매하기</div>
-                </Button>
-
-                <Button className="stock-detail-buy-button">
-                  <div className="stock-detail-buy-text">구매하기</div>
-                </Button>
-              </Row>
-            ) : (
-              <Row className="stock-detail-transaction-button">
-                <Col>
-                  <Button
-                    className="stock-detail-interest-button"
-                    onClick={handleAddToInterestStock} // 클릭 핸들러 변경
-                    disabled={isInterestStock === true}
-                  >
-                    <div className="stock-detail-interest-text">찜하기</div>
-                  </Button>
-                </Col>
-              </Row>
-            )}
-
-            {/* 모달 추가 */}
-            <ApproveInterestModal
-              isOpen={isModalOpen}
-              closeModal={() => setIsModalOpen(false)}
-              stockName={stockInfo.stockName}
-              buttonText="찜하기"
-              color="#f46060"
+            <TradeButton
+              stockBalance={stockBalance}
+              partyKey={partyKey}
+              stockKey={stockKey}
+              stockInfo={stockInfo}
             />
           </>
         )}
